+++ cmd/omaforge/main.go
@@ -18,5 +18,5 @@ func main() {
 		Commit:    commit,
 		BuildDate: buildDate,
 	}
-	os.Exit(cli.Run(os.Args[1:], os.Stdout, os.Stderr, info))
+	os.Exit(cli.Run(os.Args[1:], os.Stdin, os.Stdout, os.Stderr, info))
 }